+++ src/Composable/ComposableAdvancedSettings.jsx
@@ -6,20 +6,10 @@ export default class ComposableAdvancedSettings extends React.Component {
   static contextType = ComposableContext;
 
   render() {
-
-    const advancedFields = [{
-      name: "class",
-      label: "Class",
-    },{
-      name: "id",
-      label: "Id",
-      hint: "This can be used for anchoring purposes",
-    }]
-
     return(
       <div className="inputs">
         <p><strong>Advanced Settings</strong></p>
-        {advancedFields.map(field => (
+        { this.context.settings.advancedSettings.map(field => (
           <div className="control-group" key={field.name}>
             <label className="control-label">{field.label}</label>
             {field.hint &&
@@ -36,7 +26,7 @@ export default class ComposableAdvancedSettings extends React.Component {
           </div>
         ))}
         <div>
-          <button onClick={this.context.functions.components.closeAdvanced} className="button__primary">‹ Close advanced settings</button>
+          <button onClick={this.props.closeAdvanced} className="button__primary">‹ Close advanced settings</button>
         </div>
       </div>
     )